Extract common JAR functionality to a parent class
